backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
supabase==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
numpy==1.26.4
